Adopt new structure for source fonts
Earlier, we were using zip extraction & everything else on the
source font itself. But now that we started implementing a new
formula structure, so this commit migrates the zip extractions
to a separate concern, so it's only one place where we'll need
to change in the future.

This also extract out some non common behaviors, like: Exe, or
Zip extraction to separate helper module, so now each formula
doesn't need to include everything but only pick the one they
need for them.
