+++ lib/fontist/formulas/source_font.rb
@@ -3,45 +3,22 @@ require "zip"
 module Fontist
   module Formulas
     class SourceFont < Formulas::Base
-      def fetch
-        paths = extract_fonts.grep(/#{font_name}/i)
-        paths.empty? ? nil : paths
-      end
+      include Formulas::Helpers::ZipExtractor
 
       private
 
-      def source
-        @source ||= Fontist::Source.formulas.source_font
+      def data_node
+        @data_node ||= "source_font"
       end
 
-      def extract_fonts
-        zip_file = download_file
-        unzip_fonts(zip_file)
-      end
-
-      def unzip_fonts(file)
-        Zip.on_exists_proc = true
-
-        Array.new.tap do |fonts|
-          Zip::File.open(file) do |zip_file|
-            zip_file.glob("fonts/*.ttf").each do |entry|
-              filename = entry.name.gsub("fonts/", "")
-
-              if filename
-                font_path = fonts_path.join(filename)
-                fonts.push(font_path.to_s)
-
-                entry.extract(font_path)
-              end
+      def extract_fonts(font_names)
+        resources(data_node) do |resource|
+          zip_extract(resource, fonts_sub_dir: "fonts/") do |fonts_paths|
+            font_names.each do |font_name|
+              match_fonts(fonts_paths, font_name)
             end
           end
         end
-      end
-
-      def download_file
-        Fontist::Downloader.download(
-          source.urls.first, file_size: source.file_size.to_i, sha: source.sha
-        )
       end
     end
   end
